+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Upgrade `rubocop` dependency from `0.15.0` to `0.16.0`
 * Fix broken `--show-linters` flag
+* Ignore `BlockAlignment`, `EndAlignment`, and `IndentationWidth` Rubocop lints
 
 ## 0.3.0